requirements.txt
appdirs==1.4.3
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.4
django-allauth==0.41.0
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-filter==2.2.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-simple-email-confirmation==0.70
django-templated-mail==1.1.1
django-treebeard==4.3.1
djangorestframework==3.11.0
djangorestframework-jsonapi==3.1.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
filelock==3.0.12
httpie==2.0.0
idna==2.9
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
Markdown==3.2.1
MarkupSafe==1.1.1
mysqlclient==2.0.0.dev1
oauthlib==3.1.0
openapi-codec==1.3.2
Pygments==2.6.1
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8

